requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements_lock.txt requirements.in
#
absl-py==2.2.2
    # via -r requirements.in
aiodns==3.4.0
    # via ccxt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.6
    # via ccxt
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.75.0
    # via -r requirements.in
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
asn1crypto==1.5.1
    # via scramp
asttokens==3.0.1
    # via stack-data
asyncpg==0.30.0
    # via -r requirements.in
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via flask
ccxt==4.4.86
    # via -r requirements.in
certifi==2025.4.26
    # via
    #   ccxt
    #   httpcore
    #   httpx
    #   influxdb-client
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pycares
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   flask
    #   uvicorn
comm==0.2.3
    # via ipywidgets
contourpy==1.3.3
    # via matplotlib
cryptography==45.0.3
    # via
    #   ccxt
    #   pyjwt
cycler==0.12.1
    # via matplotlib
dateparser==1.2.2
    # via vectorbt
decorator==5.2.1
    # via ipython
dill==0.4.0
    # via vectorbt
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via anthropic
executing==2.2.1
    # via stack-data
flask==3.0.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.0
    # via -r requirements.in
fonttools==4.61.1
    # via matplotlib
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
grpcio==1.76.0
    # via -r requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   anthropic
    #   mcp
httpx-sse==0.4.3
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
imageio==2.37.2
    # via vectorbt
influxdb-client==1.48.0
    # via -r requirements.in
iniconfig==2.1.0
    # via pytest
ipython==9.9.0
    # via ipywidgets
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.8
    # via vectorbt
itsdangerous==2.2.0
    # via flask
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via flask
jiter==0.12.0
    # via anthropic
joblib==1.5.3
    # via scikit-learn
jsonschema==4.26.0
    # via mcp
jsonschema-specifications==2025.9.1
    # via jsonschema
jupyterlab-widgets==3.0.16
    # via ipywidgets
kafka-python==2.2.10
    # via -r requirements.in
kiwisolver==1.4.9
    # via matplotlib
llvmlite==0.46.0
    # via numba
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.8
    # via vectorbt
matplotlib-inline==0.2.1
    # via ipython
mcp==1.26.0
    # via -r requirements.in
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via vectorbt
narwhals==2.15.0
    # via plotly
numba==0.63.1
    # via vectorbt
numpy==2.3.5
    # via
    #   -r requirements.in
    #   contourpy
    #   imageio
    #   matplotlib
    #   numba
    #   pandas
    #   scikit-learn
    #   scipy
    #   vectorbt
packaging==25.0
    # via
    #   matplotlib
    #   plotly
    #   pytest
pandas==2.3.3
    # via
    #   -r requirements.in
    #   vectorbt
parso==0.8.5
    # via jedi
pexpect==4.9.0
    # via ipython
pg8000==1.31.2
    # via testing-postgresql
pillow==12.1.0
    # via
    #   imageio
    #   matplotlib
plotly==6.5.1
    # via vectorbt
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via ipython
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycares==4.8.0
    # via aiodns
pycparser==2.22
    # via cffi
pydantic==2.11.10
    # via
    #   anthropic
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.13.1
    # via mcp
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
pyjwt[crypto]==2.11.0
    # via mcp
pyparsing==3.3.1
    # via matplotlib
pytest==8.4.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.24.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   influxdb-client
    #   matplotlib
    #   pandas
    #   pg8000
python-dotenv==1.2.2
    # via pydantic-settings
python-multipart==0.0.22
    # via mcp
pytz==2025.2
    # via
    #   dateparser
    #   pandas
    #   vectorbt
pyyaml==6.0.3
    # via -r requirements.in
reactivex==4.0.4
    # via influxdb-client
redis==6.1.0
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via dateparser
requests==2.32.3
    # via
    #   -r requirements.in
    #   ccxt
    #   vectorbt
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
schedule==1.2.2
    # via vectorbt
scikit-learn==1.8.0
    # via vectorbt
scipy==1.16.3
    # via
    #   scikit-learn
    #   vectorbt
scramp==1.4.5
    # via pg8000
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anthropic
sse-starlette==3.3.2
    # via mcp
stack-data==0.6.3
    # via ipython
starlette==0.52.1
    # via
    #   mcp
    #   sse-starlette
tenacity==9.1.2
    # via -r requirements.in
testing-common-database==2.0.3
    # via testing-postgresql
testing-postgresql==1.3.0
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.1
    # via vectorbt
traitlets==5.14.3
    # via
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
typing-extensions==4.13.2
    # via
    #   anthropic
    #   anyio
    #   ccxt
    #   grpcio
    #   mcp
    #   pydantic
    #   pydantic-core
    #   reactivex
    #   referencing
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   mcp
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
tzlocal==5.3.1
    # via dateparser
urllib3==2.4.0
    # via
    #   influxdb-client
    #   requests
uvicorn==0.41.0
    # via mcp
vectorbt==0.28.2
    # via -r requirements.in
wcwidth==0.2.14
    # via prompt-toolkit
werkzeug==3.1.3
    # via flask
widgetsnbextension==4.0.15
    # via ipywidgets
yarl==1.20.0
    # via
    #   aiohttp
    #   ccxt

# The following packages are considered to be unsafe in a requirements file:
setuptools==82.0.0
    # via
    #   ccxt
    #   influxdb-client
